backend/test_api_endpoints.py
#!/usr/bin/env python3
"""
Test all API endpoints
"""
import requests
import json
import time

BASE_URL = "http://localhost:8000"

def test_endpoint(method, path, expected_status=200, data=None, files=None):
    """Test an API endpoint"""
    try:
        if method == "GET":
            response = requests.get(f"{BASE_URL}{path}", timeout=5)
        elif method == "POST":
            response = requests.post(f"{BASE_URL}{path}", json=data, files=files, timeout=10)
        else:
            print(f"❌ Unknown method: {method}")
            return False
        
        if response.status_code == expected_status:
            print(f"✅ {method} {path} - Status: {response.status_code}")
            if response.headers.get('content-type', '').startswith('application/json'):
                try:
                    result = response.json()
                    if isinstance(result, dict) and 'message' in result:
                        print(f"   Message: {result['message']}")
                except:
                    pass
            return True
        else:
            print(f"❌ {method} {path} - Expected {expected_status}, got {response.status_code}")
            print(f"   Response: {response.text[:200]}")
            return False
    except requests.exceptions.ConnectionError:
        print(f"❌ {method} {path} - Cannot connect to server (is it running?)")
        return False
    except Exception as e:
        print(f"❌ {method} {path} - Error: {e}")
        return False

print("🔍 Testing API Endpoints...")
print("=" * 50)

# Test basic endpoints
test_endpoint("GET", "/")
test_endpoint("GET", "/health")

# Test login endpoint (needs Form data, not JSON)
# test_endpoint("POST", "/login", expected_status=401, data={"username": "test", "password": "test"})
print("⚠️  /login endpoint requires Form data (test manually)")

# Test protected endpoint (should fail without token - returns 403, not 401)
test_endpoint("GET", "/protected", expected_status=403)

print("\n✅ Basic endpoint tests completed!")
print("\n⚠️  Note: Image upload tests require actual image files")
print("   Run these manually with curl or Postman")

