backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.26.0
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
playwright==1.41.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
pillow==10.2.0
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
opencv-python>=4.9.0
pytesseract==0.3.10
numpy>=1.24.0,<2.0.0
redis==5.0.1
celery==5.3.4
# Note: Also install Tesseract OCR system package:
# macOS: brew install tesseract
# Linux: sudo apt-get install tesseract-ocr

